App Folder Restructure
- Following suite of the main components folder refresh, fixing app folder structure
