+++ tempo-starter/src/app/(dashboard)/dashboard/page.tsx
@@ -1,7 +1,7 @@
 import DashboardNavbar from "@/components/layout/dashboard-navbar";
 import { InfoIcon, UserCircle } from "lucide-react";
 import { redirect } from "next/navigation";
-import { createClient } from "../../../supabase/server";
+import { createClient } from "../../../../supabase/server";
 
 export default async function Dashboard() {
   const supabase = await createClient();
