+++ tempo-starter/src/app/(public)/gallery/page.tsx
