+++ tempo-starter/src/app/_actions/index.ts
@@ -1,0 +1,2 @@
+// Re-export all actions for backward compatibility and clean imports
+export * from './auth';
